requirements.txt
anthropic>=0.39.0
streamlit>=1.40.0
python-docx>=1.1.0
olefile>=0.47
pypdf>=4.0.0
